requirements.txt
Twisted==16.3.0
argparse==1.2.1
nose==1.3.7
pyserial==3.1.1
wsgiref==0.1.2
zope.interface==4.2.0
